requirements.txt
appdirs==1.4.4
backports.entry-points-selectable==1.1.0
backports.weakref==1.0.post1
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
future==0.18.2
gunicorn==20.1.0
httpie==2.6.0
json5==0.9.5
mysqlclient==2.1.0
pep8==1.7.1
pipenv==2021.5.29
psycopg2==2.9.1
pycurl==7.43.0.6
sqlparse==0.4.2
statsmodels==0.12.2
tables==3.6.1
virtualenv==20.8.1
virtualenv-clone==0.5.7
whitenoise==5.3.0
PyJWT==2.3.0